requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.65.0
faiss-cpu>=1.7.4
# For GPU support, use: faiss-gpu>=1.7.4
fashion-clip>=0.2.0
scikit-learn>=1.3.0
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
transformers>=4.30.0
